requirements.txt
pandas
sshtunnel
pymysql
sqlalchemy
python-dotenv
cryptography
